refactor(datagrid-web): move filtering to the hook
